Tidy the Transitions renderer and tighten its specs, DEV-20402
Drops dead code and stale comments, folds duplicated stylesheet declarations and
splits the data shaping out of the composable. Applies one rounding rule to
every percentage, names the incoming block "Incoming traffic" when no group is
open, and stops the ribbons being invalidated on every hover. Each change now
has a spec that fails without it, and the baselines are resynced.
