+++ plugins/CustomDimensions/tests/UI/expected-screenshots/CustomDimensions_report_action_subtable_transitions.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f43a1a2500bce4eb131b997fe2e564c1b73181532bd1e4efa049e5de1e710e2b
-size 82948
+oid sha256:5d936333dbe3308ee8a8c850c22fce96302dc66fdaed6a4472c6110f61aadc8c
+size 82916
