+++ plugins/Transitions/tests/UI/expected-screenshots/Transitions_transitions_popup_titles.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:0ac467e72cea9ed092b7def2c3404793b2f3ee9a1528ee43269188d18ddf17aa
-size 105788
+oid sha256:cc05568e4ef902e687a02ca397dbe2bd75f9f5effc4d570597a003f91654dffc
+size 105082
