+++ plugins/Transitions/tests/UI/expected-screenshots/Transitions_transitions_report_dark_mode.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:b4bc3d0aa2a61e9cd10de3e6193d6932db7eef2e615133b7b21f19a1c02b98b2
-size 141049
+oid sha256:1996ecbd4b20d3aec106dfcebe187170fd5ee14bd616bc6cc85259dcc589aa67
+size 140586
