+++ plugins/Transitions/tests/UI/expected-screenshots/Transitions_transitions_report_period_not_allowed.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:a6aa1e395f18ddec33995a75ed39cad784f09fa6b950adb5ea33757e8c11aaf3
-size 143119
+oid sha256:a0f2cd9f0da01f6cd263ac0fea5e2e28a8e22fc9712a6d6995b3efbff24af3f3
+size 142420
