+++ plugins/Transitions/tests/UI/expected-screenshots/Transitions_transitions_report_search_engines.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:bdf39899190c7eb254a9e13d03b3020731b81482052ac2ce0ef07a7121fa1a8d
-size 201012
+oid sha256:7e6236e79c799adb3101b3543d58318a1d6fe252d00c1a41421a8d829f078baa
+size 200178
