+++ plugins/Transitions/tests/UI/expected-screenshots/Transitions_transitions_report_switch_type_title.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:5bc0fc6befd0bf491255947befb53e163b1cfec8dcba417aab397dddbef6c75a
-size 201450
+oid sha256:56308fe5395b69dd544c6d28c4c504d9fa9d514d7e6e2cfa6e28c0800e94a702
+size 200689
