+++ plugins/Transitions/tests/UI/expected-screenshots/Transitions_transitions_report_switch_url.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:4b2f0f8e8cca26198fad257766d08ae4c61a7babf35d7a38c2349552e73d30dc
-size 106385
+oid sha256:fa8694d2e72ffecea5cebe7984365fc92d871b2c4a954bb9d2b1da277d8b2f9d
+size 106399
