+++ plugins/Transitions/tests/UI/expected-screenshots/Transitions_transitions_report_with_data_widget.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:9ba446b0b1e012b8a5de5e7dc7190a8eec446e0801a452255f9238fd548afc85
-size 139756
+oid sha256:e63592064ae36ca6666ad1eb97dfed1d4072693e2f6ad36778836f25399902b6
+size 139051
