+++ plugins/Transitions/vue/src/TransitionsReport/TransitionsColumn.less
@@ -2,5 +2,4 @@
   display: flex;
   flex-direction: column;
   gap: 22px;
-  min-width: 0;
 }
