+++ plugins/Transitions/vue/src/TransitionsReport/ribbonGeometry.ts
@@ -15,7 +15,10 @@ const HAIRLINE_THICKNESS = 1;
 
 export interface RibbonRow {
   key: string;
-  /** Share of all pageviews, 0..1. Negative and non-finite values are treated as 0. */
+  /**
+   * Share of this page's pageviews, 0..1. Negative and non-finite values count as zero share, so
+   * the row still gets a band -- at MIN_RIBBON_THICKNESS, like any other row below the minimum.
+   */
   share: number;
   /** Row top edge, in the ribbon layer's coordinate space. */
   top: number;
@@ -31,7 +34,6 @@ export interface RibbonLayout {
   centerTop: number;
   /** Height available for the whole band on the center card edge. */
   centerHeight: number;
-  minThickness?: number;
   /** How far the row end reaches back into the row, so the join has no seam. Defaults to 0. */
   rowOverlap?: number;
   /**
@@ -56,9 +58,9 @@ function sanitiseShare(share: number): number {
  * Lays out one side's ribbons.
  *
  * Each row gets a share of the center band proportional to its share of the column total, clamped
- * up to `minThickness` so small rows stay visible. Clamped rows keep that minimum and the rest
- * share what is left of the band, so the minimum is not scaled away again. Only when the minimums
- * alone overflow the band does everything scale down (never below a hairline).
+ * up to MIN_RIBBON_THICKNESS so small rows stay visible. Clamped rows keep that minimum and the
+ * rest share what is left of the band, so the minimum is not scaled away again. Only when the
+ * minimums alone overflow the band does everything scale down (never below a hairline).
  *
  * Ribbons always run from the row to the center edge: for the incoming side that is left to
  * right, for the outgoing side right to left. The row end can reach back past the layer's edge and
@@ -69,7 +71,6 @@ function sanitiseShare(share: number): number {
  */
 export function computeRibbonPaths(rows: RibbonRow[], layout: RibbonLayout): RibbonPath[] {
   const { side, width, centerTop, centerHeight } = layout;
-  const minThickness = layout.minThickness ?? MIN_RIBBON_THICKNESS;
 
   if (!rows.length || width <= 0 || centerHeight <= 0) {
     return [];
@@ -81,9 +82,9 @@ export function computeRibbonPaths(rows: RibbonRow[], layout: RibbonLayout): Rib
   }
 
   const raw = rows.map((row) => (sanitiseShare(row.share) / total) * centerHeight);
-  const isClamped = raw.map((thickness) => thickness < minThickness);
+  const isClamped = raw.map((thickness) => thickness < MIN_RIBBON_THICKNESS);
 
-  const clampedTotal = isClamped.filter(Boolean).length * minThickness;
+  const clampedTotal = isClamped.filter(Boolean).length * MIN_RIBBON_THICKNESS;
   const unclampedTotal = raw.reduce(
     (sum, thickness, index) => (isClamped[index] ? sum : sum + thickness),
     0,
@@ -96,13 +97,13 @@ export function computeRibbonPaths(rows: RibbonRow[], layout: RibbonLayout): Rib
     const scale = centerHeight / (clampedTotal + unclampedTotal);
     thicknesses = raw.map((thickness, index) => Math.max(
       HAIRLINE_THICKNESS,
-      (isClamped[index] ? minThickness : thickness) * scale,
+      (isClamped[index] ? MIN_RIBBON_THICKNESS : thickness) * scale,
     ));
   } else {
     // Clamped rows keep their minimum; the rest share what is left, in proportion.
     const scale = budget / unclampedTotal;
     thicknesses = raw.map(
-      (thickness, index) => (isClamped[index] ? minThickness : thickness * scale),
+      (thickness, index) => (isClamped[index] ? MIN_RIBBON_THICKNESS : thickness * scale),
     );
   }
