+++ plugins/Transitions/vue/src/TransitionsReport/testTransitionsReportHarness.ts
@@ -1,0 +1,38 @@
+/*!
+ * Matomo - free/libre analytics platform
+ *
+ * @link    https://matomo.org
+ * @license https://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
+ */
+
+import { mount, VueWrapper } from '@vue/test-utils';
+import { nextTick } from 'vue';
+import TransitionsReport from './TransitionsReport.vue';
+
+/** Mounts the report with the sanitize helpers Matomo installs globally. */
+export function mountTransitionsReport(props = {}): VueWrapper {
+  return mount(TransitionsReport as any, { // eslint-disable-line @typescript-eslint/no-explicit-any
+    props: {
+      actionType: 'url',
+      actionName: 'http://example.org/page',
+      ...props,
+    },
+    global: {
+      config: {
+        globalProperties: {
+          $sanitize: (value: string) => value,
+          // stands in for DOMPurify.isValidAttribute: rejects anything not http(s)
+          $sanitizeUrl: (url: string) => (/^https?:\/\//i.test(url) ? url : ''),
+        },
+      },
+    },
+  });
+}
+
+/** The grid, then the resolved element refs, then the ribbon paths each take a render. */
+export async function flushRibbons(): Promise<void> {
+  for (let i = 0; i < 5; i += 1) {
+    // eslint-disable-next-line no-await-in-loop
+    await nextTick();
+  }
+}
